IOS-6261: Add APIListUtils, update tests
